4th Semester/Programming for Artificial Intellegence (Lab)/Assignment 6/requirements.txt
Pillow
scikit-learn
numpy
opencv-python
Flask